geoclip_api/requirements.txt
streamlit>=1.24.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
requests>=2.31.0
Pillow>=9.5.0
geopy>=2.3.0
nest-asyncio>=1.5.6
torch>=2.6.0
numpy>=1.24.0
geoclip>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.2.0
prometheus-client>=0.17.0
transformers>=4.30.0
accelerate>=0.20.0 